requirements.txt
streamlit
numpy
pillow
scikit-learn
joblib
pandas
